+++ tcvectordb/consts.go
@@ -25,16 +25,17 @@ type IndexType string
 
 const (
 	// vector index type
-	FLAT      IndexType = "FLAT"
-	HNSW      IndexType = "HNSW"
-	IVF_FLAT  IndexType = "IVF_FLAT"
-	IVF_PQ    IndexType = "IVF_PQ"
-	IVF_SQ4   IndexType = "IVF_SQ4"
-	IVF_SQ8   IndexType = "IVF_SQ8"
-	IVF_SQ16  IndexType = "IVF_SQ16"
-	BIN_FLAT  IndexType = "BIN_FLAT"
-	BIN_HNSW  IndexType = "BIN_HNSW"
-	DISK_FLAT IndexType = "DISK_FLAT"
+	FLAT       IndexType = "FLAT"
+	HNSW       IndexType = "HNSW"
+	IVF_FLAT   IndexType = "IVF_FLAT"
+	IVF_PQ     IndexType = "IVF_PQ"
+	IVF_SQ4    IndexType = "IVF_SQ4"
+	IVF_SQ8    IndexType = "IVF_SQ8"
+	IVF_SQ16   IndexType = "IVF_SQ16"
+	IVF_RABITQ IndexType = "IVF_RABITQ"
+	BIN_FLAT   IndexType = "BIN_FLAT"
+	BIN_HNSW   IndexType = "BIN_HNSW"
+	DISK_FLAT  IndexType = "DISK_FLAT"
 
 	// scalar index type
 	PRIMARY         IndexType = "primaryKey"
@@ -59,6 +60,7 @@ const (
 	Array          FieldType = "array"
 	Json           FieldType = "json"
 	Double         FieldType = "double"
+	Int64          FieldType = "int64"
 	Vector         FieldType = "vector"
 	Float16Vector  FieldType = "float16_vector"
 	BFloat16Vector FieldType = "bfloat16_vector"